websockets/interfaces/WebSocket/send/005.html

<!doctype html>
<title>WebSockets: send() return value</title>

<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../../../constants.js></script>
<div id=log></div>
<script>
async_test(function(t){
  var ws = new WebSocket(SCHEME_DOMAIN_PORT+'/ws/echo');
  ws.onopen = t.step_func(function(e) {
    assert_equals(ws.send('test'), undefined);
    t.done();
  });
});
</script>
